+++ packages/library-base/src/jvm/kotlin/io/realm/kotlin/internal/platform/SystemUtils.kt
@@ -1,8 +1,14 @@
 package io.realm.kotlin.internal.platform
 
+import io.realm.kotlin.internal.Constants.FILE_COPY_BUFFER_SIZE
 import io.realm.kotlin.internal.RealmInstantImpl
+import io.realm.kotlin.internal.util.Exceptions
 import io.realm.kotlin.types.RealmInstant
 import java.io.File
+import java.io.InputStream
+import java.io.OutputStream
+import java.security.DigestInputStream
+import java.security.MessageDigest
 import java.time.Clock.systemUTC
 import java.util.concurrent.TimeUnit
 import kotlin.reflect.KMutableProperty1
@@ -44,6 +50,61 @@ public actual fun prepareRealmDirectoryPath(directoryPath: String): String {
 public actual fun prepareRealmFilePath(directoryPath: String, filename: String): String {
     preparePath(directoryPath)
     return File(directoryPath, filename).absolutePath
+}
+
+private fun copyStream(inputStream: InputStream, outputStream: OutputStream) {
+    val buf = ByteArray(FILE_COPY_BUFFER_SIZE)
+    var bytesRead: Int
+    while ((inputStream.read(buf).also { bytesRead = it }) > -1) {
+        outputStream.write(buf, 0, bytesRead)
+    }
+}
+
+/**
+ * Open an input stream from the asset file according to the platform conventions.
+ * - Android: Through android.content.res.AssetManager.open(assetFilename)
+ * - JVM: Class<T>.javaClass.classLoader.getResource(assetFilename)
+ *
+ * @throws Exceptions.assetFileNotFound if the file is not found.
+ */
+public expect fun assetFileAsStream(assetFilename: String): InputStream
+
+@Suppress("NestedBlockDepth")
+public actual fun copyAssetFile(
+    realmFilePath: String,
+    assetFilename: String,
+    sha256Checksum: String?
+) {
+    assetFileAsStream(assetFilename).let { inputStream ->
+        if (sha256Checksum != null) {
+            DigestInputStream(inputStream, MessageDigest.getInstance("SHA-256"))
+        } else {
+            inputStream
+        }
+    }.use { assetStream ->
+        val outputFile = File(realmFilePath)
+        try {
+            outputFile.outputStream().use { outputStream ->
+                copyStream(assetStream, outputStream)
+            }
+            if (sha256Checksum != null && assetStream is DigestInputStream) {
+                val actual = assetStream.messageDigest.digest()
+                    .fold("", { str, element -> str + "%02x".format(element) })
+                if (actual != sha256Checksum) {
+                    throw Exceptions.assetFileChecksumMismatch(
+                        assetFilename,
+                        sha256Checksum,
+                        actual
+                    )
+                }
+            }
+        } catch (e: Exception) {
+            if (outputFile.exists()) {
+                outputFile.delete()
+            }
+            throw e
+        }
+    }
 }
 
 public actual fun <K : Any?, V : Any?> returnType(field: KMutableProperty1<K, V>): KType {